+++ kitty/config_data.py
@@ -325,7 +325,8 @@ Note that this code is indexed by PostScript name, and not the font
 family. This allows you to define very precise feature settings; e.g. you can
 disable a feature in the italic font but not in the regular font.
 
-By default they are derived automatically, by the OSes font system (linux only).
+On Linux, these are read from the FontConfig database first and then this,
+setting is applied, so they can be configured in a single, central place.
 
 To get the PostScript name for a font, use :code:`kitty + list-fonts --psnames`: